Fix clearRelationships on hasMany relationships when record is deleted
When a record with hasMany relationships is deleted, we need to clear out all of
its relations. This requires clearing the array hasMany relations are stored in.
The code to do this references `relationship.name` which is a nonexistent
property, so nothing ever gets cleared. The intent is to reference `name`
directly.